requirements.txt
aiogram
asyncpg
pandas
openpyxl
python-dotenv
